requirements.txt
streamlit
PyPDF2
cohere
pinecone-client
python-dotenv
IPython
ipykernel